+++ custom-view/lib-custom-view/src/main/java/ru/surfstudio/android/custom/view/TitleSubtitleView.kt
@@ -28,6 +28,7 @@ import android.widget.TextView
 import androidx.annotation.DrawableRes
 import androidx.annotation.StyleRes
 import androidx.core.content.ContextCompat.getDrawable
+import ru.surfstudio.android.logger.Logger
 
 private const val DEFAULT_MAX_LINES: Int = 1
 
@@ -223,6 +224,18 @@ class TitleSubtitleView @JvmOverloads constructor(
     }
 
     /**
+     * Устанавливает текст и тип текста заголовка
+     */
+    fun setTitleText(text: CharSequence, type: TextView.BufferType) =
+            titleView.setText(text, type)
+
+    /**
+     * Устанавливает текст и тип текста подзаголовка
+     */
+    fun setSubTitleText(text: CharSequence, type: TextView.BufferType) =
+            subTitleView.setText(text, type)
+
+    /**
      * Возвращает заголовок к дефолтному значению
      */
     fun resetTitleText() {
@@ -374,7 +387,7 @@ class TitleSubtitleView @JvmOverloads constructor(
                 2 -> TextUtils.TruncateAt.MIDDLE
                 3 -> TextUtils.TruncateAt.END
                 4 -> TextUtils.TruncateAt.MARQUEE
-                else -> TextUtils.TruncateAt.END
+                else -> null
             }
 
     private fun getVisibilityFromResource(ta: TypedArray, index: Int): Int =